% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_Permission.R
\name{add_permission}
\alias{add_permission}
\title{add permission}
\usage{
add_permission(name, permission)
}
\arguments{
\item{name}{the name of the permission}

\item{permission}{a Permission R6 class}
}
\description{
add permission
}
