% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v1_add_schema.r
\name{add_schema}
\alias{add_schema}
\title{add schema}
\usage{
add_schema(name = NULL, schema)
}
\arguments{
\item{name}{the name of the schema}

\item{schema}{a Schema R6 class}
}
\description{
add schema
}
