% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v8_add_schema.R
\name{add_schema_v8}
\alias{add_schema_v8}
\title{add schema}
\usage{
add_schema_v8(
  name_access = NULL,
  name_grouping = NULL,
  name_variant = NULL,
  db_configs = NULL,
  field_types,
  keys,
  censors = NULL,
  indexes = NULL,
  validator_field_types = validator_field_types_blank,
  validator_field_contents = validator_field_contents_blank,
  info = NULL
)
}
\arguments{
\item{name_access}{a}

\item{name_grouping}{a}

\item{name_variant}{a}

\item{db_configs}{a}

\item{field_types}{a}

\item{keys}{a}

\item{censors}{a}

\item{indexes}{a}

\item{validator_field_types}{a}

\item{validator_field_contents}{a}

\item{info}{a}
}
\description{
add schema
}
