% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v8_add_task_from_config.r
\name{add_task_from_config_v8}
\alias{add_task_from_config_v8}
\title{add_task_from_config_v8}
\usage{
add_task_from_config_v8(
  name = NULL,
  name_grouping = NULL,
  name_action = NULL,
  name_variant = NULL,
  cores = 1,
  permission = NULL,
  plan_analysis_fn_name = NULL,
  for_each_plan = NULL,
  for_each_analysis = NULL,
  universal_argset = NULL,
  upsert_at_end_of_each_plan = FALSE,
  insert_at_end_of_each_plan = FALSE,
  action_fn_name,
  data_selector_fn_name = NULL,
  schema = NULL,
  info = NULL
)
}
\arguments{
\item{name}{Name of the task (dont use)}

\item{name_grouping}{Name of the task (grouping)}

\item{name_action}{Name of the task (action)}

\item{name_variant}{Name of the task (variant)}

\item{cores}{Number of CPU cores}

\item{permission}{A permission R6 instance}

\item{plan_analysis_fn_name}{The name of a function that returns a named list \code{list(for_each_plan = list(), for_each_analysis = NULL)}.}

\item{for_each_plan}{A list, where each unit corresponds to one data extraction. Generally recommended to use \code{plnr::expand_list}.}

\item{for_each_analysis}{A list, where each unit corresponds to one analysis within a plan (data extraction). Generally recommended to use \code{plnr::expand_list}.}

\item{universal_argset}{A list, where these argsets are applied to all analyses univerally}

\item{upsert_at_end_of_each_plan}{Do you want to upsert your results automatically at the end of each plan?}

\item{insert_at_end_of_each_plan}{Do you want to insert your results automatically at the end of each plan?}

\item{action_fn_name}{The name of the function that will be called for each analysis with arguments \code{data}, \code{argset}, \code{schema}}

\item{data_selector_fn_name}{The name of a function that will be called to obtain the data for each analysis. The function must have the arguments \code{argset}, \code{schema} and must return a named list.}

\item{schema}{A named list that maps \code{sc8::config$schemas} for use in \code{action_fn_name} and \code{data_selector_fn_name}}

\item{info}{Information for documentation}
}
\description{
add_task_from_config_v8
}
