% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_SchemaRedirect.r
\name{censor_list_function_factory}
\alias{censor_list_function_factory}
\title{censor 0-4 function factory}
\usage{
censor_list_function_factory(list_of_censors)
}
\arguments{
\item{list_of_censors}{List of censors}
}
\description{
censor 0-4 function factory
}
