% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_database.R
\name{drop_all_rows}
\alias{drop_all_rows}
\title{drop_all_rows
Drops all rows}
\usage{
drop_all_rows(conn = NULL, table)
}
\arguments{
\item{conn}{A db connection}

\item{table}{Table name}
}
\description{
drop_all_rows
Drops all rows
}
