% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_database.R
\name{drop_rows_where}
\alias{drop_rows_where}
\title{Drops the rows where the condition is met}
\usage{
drop_rows_where(conn = NULL, table, condition)
}
\arguments{
\item{conn}{A db connection}

\item{table}{Table name}

\item{condition}{A string SQL condition}
}
\description{
Drops the rows where the condition is met
}
