% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v8_fill_in_missing.r
\name{fill_in_missing_v8}
\alias{fill_in_missing_v8}
\title{fill_in_missing_v8}
\usage{
fill_in_missing_v8(d, border = 2020)
}
\arguments{
\item{d}{dataset}

\item{border}{what border?}
}
\description{
fill_in_missing_v8
}
