% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_config_last_updated.r
\name{get_config_last_updated}
\alias{get_config_last_updated}
\title{get_config_last_updated
Gets the config_last_updated db table}
\usage{
get_config_last_updated(type = NULL, tag = NULL)
}
\arguments{
\item{type}{a}

\item{tag}{a}
}
\description{
get_config_last_updated
Gets the config_last_updated db table
}
