% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_Schema.R
\name{identify_data_structure.Schema_v8}
\alias{identify_data_structure.Schema_v8}
\title{Hashing data structure (Schema_v8)}
\usage{
\method{identify_data_structure}{Schema_v8}(x, col, ...)
}
\arguments{
\item{x}{A Schema_v8 object}

\item{col}{The column to be hashed}

\item{...}{Unused}
}
\description{
An implementation of spltidy::identify_data_structure for Schema_v8
}
