% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_database.R
\name{keep_rows_where}
\alias{keep_rows_where}
\title{keep_rows_where
Keeps the rows where the condition is met}
\usage{
keep_rows_where(conn = NULL, table, condition)
}
\arguments{
\item{conn}{A db connection}

\item{table}{Table name}

\item{condition}{A string SQL condition}
}
\description{
keep_rows_where
Keeps the rows where the condition is met
}
