% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_Permission.R
\name{Permission}
\alias{Permission}
\title{Permission}
\description{
Permission

Permission
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Permission$new()}}
\item \href{#method-has_permission}{\code{Permission$has_permission()}}
\item \href{#method-revoke_permission}{\code{Permission$revoke_permission()}}
\item \href{#method-grant_permission}{\code{Permission$grant_permission()}}
\item \href{#method-is_final}{\code{Permission$is_final()}}
\item \href{#method-current_value}{\code{Permission$current_value()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Permission$new(key, value, production_days = c(1:7))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-has_permission"></a>}}
\if{latex}{\out{\hypertarget{method-has_permission}{}}}
\subsection{Method \code{has_permission()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Permission$has_permission()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-revoke_permission"></a>}}
\if{latex}{\out{\hypertarget{method-revoke_permission}{}}}
\subsection{Method \code{revoke_permission()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Permission$revoke_permission()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grant_permission"></a>}}
\if{latex}{\out{\hypertarget{method-grant_permission}{}}}
\subsection{Method \code{grant_permission()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Permission$grant_permission()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_final"></a>}}
\if{latex}{\out{\hypertarget{method-is_final}{}}}
\subsection{Method \code{is_final()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Permission$is_final()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-current_value"></a>}}
\if{latex}{\out{\hypertarget{method-current_value}{}}}
\subsection{Method \code{current_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Permission$current_value()}\if{html}{\out{</div>}}
}

}
}
