% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_env.R
\docType{data}
\name{pools}
\alias{pools}
\title{Flags/values to be used in the 'dashboards' scene}
\format{
An object of class \code{environment} of length 4.
}
\usage{
pools
}
\description{
Flags/values to be used in the 'dashboards' scene
}
\keyword{datasets}
