% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_enc.R
\name{qsenc_read}
\alias{qsenc_read}
\title{Reads an object in an encrypted file serialized to disk}
\usage{
qsenc_read(file, private_key_path = Sys.getenv("ENCRYPTR_ID_RSA"))
}
\arguments{
\item{file}{The file name/path (must end in '.qs.enc')}

\item{private_key_path}{Path to private key}
}
\description{
Reads an object in an encrypted file serialized to disk
}
