% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_enc.R
\name{qsenc_save}
\alias{qsenc_save}
\title{Saves (serializes) an encrypted object to disk}
\usage{
qsenc_save(
  x,
  file,
  nthreads = 1,
  public_key_path = Sys.getenv("ENCRYPTR_ID_RSA_PUB")
)
}
\arguments{
\item{x}{The object to serialize}

\item{file}{The file name/path (must end in '.qs.enc')}

\item{nthreads}{Number of threads to use. Default 1.}

\item{public_key_path}{Path to public key}
}
\description{
Saves (serializes) an encrypted object to disk
}
