% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_SurveillanceSystem.R
\name{run_task_sequentially_as_rstudio_job_loading_from_devtools}
\alias{run_task_sequentially_as_rstudio_job_loading_from_devtools}
\title{Run}
\usage{
run_task_sequentially_as_rstudio_job_loading_from_devtools(
  task_name,
  ss_prefix = "global$ss"
)
}
\arguments{
\item{task_name}{Task name}

\item{ss_prefix}{The prefix that locates the surveillance system}
}
\description{
Run
}
