% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_SurveillanceSystem.R
\name{run_task_sequentially_as_rstudio_job_using_load_all}
\alias{run_task_sequentially_as_rstudio_job_using_load_all}
\title{Run a Task Sequentially as an RStudio Job}
\usage{
run_task_sequentially_as_rstudio_job_using_load_all(
  task_name,
  ss_prefix = "global$ss"
)
}
\arguments{
\item{task_name}{Character string specifying the name of the task to run.}

\item{ss_prefix}{Character string specifying the prefix used to access the
surveillance system object. Defaults to "global$ss".}
}
\value{
No return value. This function is called for its side effect of
launching an RStudio job that executes the surveillance task.
}
\description{
Executes a surveillance task as an RStudio job using devtools::load_all()
for package development. This function creates a temporary R script that
loads the package and runs the specified task sequentially (cores = 1).
}
\details{
Use this function mainly during package development, to test tasks
interactively. It creates a temporary R script that:
\itemize{
\item Loads the package using \code{devtools::load_all()}.
\item Sets the task to run with single core (cores = 1).
\item Executes the task via the surveillance system.
}
}
\examples{
\dontrun{
# Run a task as RStudio job during development
run_task_sequentially_as_rstudio_job_using_load_all("covid_analysis")

# Use custom surveillance system prefix
run_task_sequentially_as_rstudio_job_using_load_all(
  "covid_analysis", 
  ss_prefix = "my_ss"
)
}

}
