% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_test.r
\name{sc_test_action}
\alias{sc_test_action}
\title{sc_test_action (action)}
\usage{
sc_test_action(data, argset, schema)
}
\arguments{
\item{data}{Data}

\item{argset}{Argset}

\item{schema}{DB Schema}
}
\description{
sc_test_action (action)
}
