% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_test.r
\name{sc_test_data_selector}
\alias{sc_test_data_selector}
\title{sc_test (data selector)}
\usage{
sc_test_data_selector(argset, schema)
}
\arguments{
\item{argset}{Argset}

\item{schema}{DB Schema}
}
\description{
sc_test (data selector)
}
