% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_test.r
\name{sc_test_plan_analysis}
\alias{sc_test_plan_analysis}
\title{sc_test (plan/analysis)
This function can be deleted if you are not using "plan_analysis_fn_name"
inside sc8::task_from_config_v3}
\usage{
sc_test_plan_analysis(argset, schema)
}
\arguments{
\item{argset}{argset}

\item{schema}{schema}
}
\description{
sc_test (plan/analysis)
This function can be deleted if you are not using "plan_analysis_fn_name"
inside sc8::task_from_config_v3
}
