% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_Schema.R
\name{Schema}
\alias{Schema}
\title{schema class description}
\description{
schema class description

schema class description
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-load_folder_fn}{\code{Schema$load_folder_fn()}}
\item \href{#method-new}{\code{Schema$new()}}
\item \href{#method-connect}{\code{Schema$connect()}}
\item \href{#method-print}{\code{Schema$print()}}
\item \href{#method-disconnect}{\code{Schema$disconnect()}}
\item \href{#method-create_table}{\code{Schema$create_table()}}
\item \href{#method-drop_table}{\code{Schema$drop_table()}}
\item \href{#method-insert_data}{\code{Schema$insert_data()}}
\item \href{#method-upsert_data}{\code{Schema$upsert_data()}}
\item \href{#method-drop_all_rows}{\code{Schema$drop_all_rows()}}
\item \href{#method-drop_rows_where}{\code{Schema$drop_rows_where()}}
\item \href{#method-keep_rows_where}{\code{Schema$keep_rows_where()}}
\item \href{#method-drop_all_rows_and_then_upsert_data}{\code{Schema$drop_all_rows_and_then_upsert_data()}}
\item \href{#method-drop_all_rows_and_then_insert_data}{\code{Schema$drop_all_rows_and_then_insert_data()}}
\item \href{#method-tbl}{\code{Schema$tbl()}}
\item \href{#method-db_connect}{\code{Schema$db_connect()}}
\item \href{#method-db_disconnect}{\code{Schema$db_disconnect()}}
\item \href{#method-db_add_constraint}{\code{Schema$db_add_constraint()}}
\item \href{#method-db_drop_constraint}{\code{Schema$db_drop_constraint()}}
\item \href{#method-db_create_table}{\code{Schema$db_create_table()}}
\item \href{#method-db_drop_table}{\code{Schema$db_drop_table()}}
\item \href{#method-db_check_fields_match}{\code{Schema$db_check_fields_match()}}
\item \href{#method-db_insert_data}{\code{Schema$db_insert_data()}}
\item \href{#method-db_upsert_data}{\code{Schema$db_upsert_data()}}
\item \href{#method-db_load_data_infile}{\code{Schema$db_load_data_infile()}}
\item \href{#method-db_upsert_load_data_infile}{\code{Schema$db_upsert_load_data_infile()}}
\item \href{#method-db_drop_all_rows}{\code{Schema$db_drop_all_rows()}}
\item \href{#method-db_drop_rows_where}{\code{Schema$db_drop_rows_where()}}
\item \href{#method-db_keep_rows_where}{\code{Schema$db_keep_rows_where()}}
\item \href{#method-db_drop_all_rows_and_then_upsert_data}{\code{Schema$db_drop_all_rows_and_then_upsert_data()}}
\item \href{#method-get_data}{\code{Schema$get_data()}}
\item \href{#method-get_data_dt}{\code{Schema$get_data_dt()}}
\item \href{#method-get_data_db}{\code{Schema$get_data_db()}}
\item \href{#method-dplyr_tbl}{\code{Schema$dplyr_tbl()}}
\item \href{#method-list_indexes_db}{\code{Schema$list_indexes_db()}}
\item \href{#method-add_indexes}{\code{Schema$add_indexes()}}
\item \href{#method-drop_indexes}{\code{Schema$drop_indexes()}}
\item \href{#method-identify_dt_that_exists_in_db}{\code{Schema$identify_dt_that_exists_in_db()}}
\item \href{#method-clone}{\code{Schema$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_folder_fn"></a>}}
\if{latex}{\out{\hypertarget{method-load_folder_fn}{}}}
\subsection{Method \code{load_folder_fn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$load_folder_fn()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$new(
  dt = NULL,
  conn = NULL,
  db_config = NULL,
  db_table,
  db_field_types,
  db_load_folder,
  keys,
  indexes = NULL,
  validator_field_types = validator_field_types_blank,
  validator_field_contents = validator_field_contents_blank,
  info = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connect"></a>}}
\if{latex}{\out{\hypertarget{method-connect}{}}}
\subsection{Method \code{connect()}}{
Connect to a db
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$connect(db_config = self$db_config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db_config}}{db_config}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-disconnect}{}}}
\subsection{Method \code{disconnect()}}{
Disconnect from a db
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$disconnect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_table"></a>}}
\if{latex}{\out{\hypertarget{method-create_table}{}}}
\subsection{Method \code{create_table()}}{
Create db table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$create_table()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-drop_table"></a>}}
\if{latex}{\out{\hypertarget{method-drop_table}{}}}
\subsection{Method \code{drop_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$drop_table()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-insert_data"></a>}}
\if{latex}{\out{\hypertarget{method-insert_data}{}}}
\subsection{Method \code{insert_data()}}{
Inserts data into db table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$insert_data(newdata, verbose = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upsert_data"></a>}}
\if{latex}{\out{\hypertarget{method-upsert_data}{}}}
\subsection{Method \code{upsert_data()}}{
Upserts data into db table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$upsert_data(newdata, drop_indexes = names(self$indexes), verbose = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-drop_all_rows"></a>}}
\if{latex}{\out{\hypertarget{method-drop_all_rows}{}}}
\subsection{Method \code{drop_all_rows()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$drop_all_rows()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-drop_rows_where"></a>}}
\if{latex}{\out{\hypertarget{method-drop_rows_where}{}}}
\subsection{Method \code{drop_rows_where()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$drop_rows_where(condition)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-keep_rows_where"></a>}}
\if{latex}{\out{\hypertarget{method-keep_rows_where}{}}}
\subsection{Method \code{keep_rows_where()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$keep_rows_where(condition)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-drop_all_rows_and_then_upsert_data"></a>}}
\if{latex}{\out{\hypertarget{method-drop_all_rows_and_then_upsert_data}{}}}
\subsection{Method \code{drop_all_rows_and_then_upsert_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$drop_all_rows_and_then_upsert_data(
  newdata,
  drop_indexes = names(self$indexes),
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-drop_all_rows_and_then_insert_data"></a>}}
\if{latex}{\out{\hypertarget{method-drop_all_rows_and_then_insert_data}{}}}
\subsection{Method \code{drop_all_rows_and_then_insert_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$drop_all_rows_and_then_insert_data(newdata, verbose = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tbl"></a>}}
\if{latex}{\out{\hypertarget{method-tbl}{}}}
\subsection{Method \code{tbl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$tbl()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_connect"></a>}}
\if{latex}{\out{\hypertarget{method-db_connect}{}}}
\subsection{Method \code{db_connect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$db_connect(db_config = self$db_config)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-db_disconnect}{}}}
\subsection{Method \code{db_disconnect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$db_disconnect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_add_constraint"></a>}}
\if{latex}{\out{\hypertarget{method-db_add_constraint}{}}}
\subsection{Method \code{db_add_constraint()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$db_add_constraint()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_drop_constraint"></a>}}
\if{latex}{\out{\hypertarget{method-db_drop_constraint}{}}}
\subsection{Method \code{db_drop_constraint()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$db_drop_constraint()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_create_table"></a>}}
\if{latex}{\out{\hypertarget{method-db_create_table}{}}}
\subsection{Method \code{db_create_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$db_create_table()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_drop_table"></a>}}
\if{latex}{\out{\hypertarget{method-db_drop_table}{}}}
\subsection{Method \code{db_drop_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$db_drop_table()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_check_fields_match"></a>}}
\if{latex}{\out{\hypertarget{method-db_check_fields_match}{}}}
\subsection{Method \code{db_check_fields_match()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$db_check_fields_match()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_insert_data"></a>}}
\if{latex}{\out{\hypertarget{method-db_insert_data}{}}}
\subsection{Method \code{db_insert_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$db_insert_data(newdata, verbose = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_upsert_data"></a>}}
\if{latex}{\out{\hypertarget{method-db_upsert_data}{}}}
\subsection{Method \code{db_upsert_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$db_upsert_data(
  newdata,
  drop_indexes = names(self$indexes),
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_load_data_infile"></a>}}
\if{latex}{\out{\hypertarget{method-db_load_data_infile}{}}}
\subsection{Method \code{db_load_data_infile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$db_load_data_infile(newdata, verbose = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_upsert_load_data_infile"></a>}}
\if{latex}{\out{\hypertarget{method-db_upsert_load_data_infile}{}}}
\subsection{Method \code{db_upsert_load_data_infile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$db_upsert_load_data_infile(newdata, verbose = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_drop_all_rows"></a>}}
\if{latex}{\out{\hypertarget{method-db_drop_all_rows}{}}}
\subsection{Method \code{db_drop_all_rows()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$db_drop_all_rows()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_drop_rows_where"></a>}}
\if{latex}{\out{\hypertarget{method-db_drop_rows_where}{}}}
\subsection{Method \code{db_drop_rows_where()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$db_drop_rows_where(condition)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_keep_rows_where"></a>}}
\if{latex}{\out{\hypertarget{method-db_keep_rows_where}{}}}
\subsection{Method \code{db_keep_rows_where()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$db_keep_rows_where(condition)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_drop_all_rows_and_then_upsert_data"></a>}}
\if{latex}{\out{\hypertarget{method-db_drop_all_rows_and_then_upsert_data}{}}}
\subsection{Method \code{db_drop_all_rows_and_then_upsert_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$db_drop_all_rows_and_then_upsert_data(
  newdata,
  drop_indexes = names(self$indexes),
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_data}{}}}
\subsection{Method \code{get_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$get_data(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data_dt"></a>}}
\if{latex}{\out{\hypertarget{method-get_data_dt}{}}}
\subsection{Method \code{get_data_dt()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$get_data_dt(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data_db"></a>}}
\if{latex}{\out{\hypertarget{method-get_data_db}{}}}
\subsection{Method \code{get_data_db()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$get_data_db(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dplyr_tbl"></a>}}
\if{latex}{\out{\hypertarget{method-dplyr_tbl}{}}}
\subsection{Method \code{dplyr_tbl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$dplyr_tbl()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_indexes_db"></a>}}
\if{latex}{\out{\hypertarget{method-list_indexes_db}{}}}
\subsection{Method \code{list_indexes_db()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$list_indexes_db()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_indexes"></a>}}
\if{latex}{\out{\hypertarget{method-add_indexes}{}}}
\subsection{Method \code{add_indexes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$add_indexes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-drop_indexes"></a>}}
\if{latex}{\out{\hypertarget{method-drop_indexes}{}}}
\subsection{Method \code{drop_indexes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$drop_indexes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-identify_dt_that_exists_in_db"></a>}}
\if{latex}{\out{\hypertarget{method-identify_dt_that_exists_in_db}{}}}
\subsection{Method \code{identify_dt_that_exists_in_db()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$identify_dt_that_exists_in_db()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Schema$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
