% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v1_task_from_config.R
\name{task_from_config}
\alias{task_from_config}
\title{task_from_config}
\usage{
task_from_config(
  name,
  type,
  cores = 1,
  db_table = NULL,
  filter = "",
  for_each_plan = NULL,
  for_each_argset = NULL,
  upsert_at_end_of_each_plan = FALSE,
  insert_at_end_of_each_plan = FALSE,
  action,
  schema = NULL,
  args = NULL
)
}
\arguments{
\item{name}{Name of task}

\item{type}{Type of task ("data","single", "analysis", "ui")}

\item{cores}{Number of cores to run the task on}

\item{db_table}{Database table}

\item{filter}{Filter for the database table}

\item{for_each_plan}{Create a plan for each value}

\item{for_each_argset}{Create an argset for each value}

\item{upsert_at_end_of_each_plan}{If TRUE, then schema$output is used to upsert at the end of each plan}

\item{insert_at_end_of_each_plan}{If TRUE, then schema$output is used to insert at the end of each plan}

\item{action}{Text}

\item{schema}{List of schema mappings}

\item{args}{List of args}
}
\description{
task_from_config
}
