% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_inline_v1.R
\name{task_inline_v1}
\alias{task_inline_v1}
\title{Creating an inline task}
\usage{
task_inline_v1(name, action_fn = function(data, argset, schema) {
    print(1)
})
}
\arguments{
\item{name}{Name of the task}

\item{action_fn}{A function with arguments \code{schema}, \code{argset} that is the default function to extract data for each unit in data_selector_schemas not covered by \code{data_selection_fn_specific}}
}
\description{
This function is used to create a task when
you need to do a very simple task, very quickly
}
\examples{
\dontrun{
task_inline_v1(
  # name of the task
  name = "ui_normomo_attrib_excel",
  # inline function that will run
  action_fn = function(data, argset, schema) {
    print(1)
  }
)
}
}
