% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_TaskManager.R
\name{TaskManager}
\alias{TaskManager}
\title{TaskManager}
\description{
An R6 Action class contains a function called 'run' that takes three arguments:
\itemize{
\item data (list)
\item arg (list)
\item schema (list)
}
}
\details{
An action is (note: we dont explicitly create these):
\itemize{
\item one R6 Action class
\item A plnr::Plan that provide
a) data (from the plan -- \code{get_data})
b) arguments (from the plan -- \code{get_argset})
to the R6 action class
}

A task is:
\itemize{
\item one R6 Action class
\item A plnr::Plan that provide
a) data (from the plan -- \code{get_data})
b) arguments (from the plan -- \code{get_argset})
to the R6 action class
}

A TaskManager is:
\itemize{
\item one R6 Action class
\item A list of plnr::Plan's
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TaskManager$new()}}
\item \href{#method-add_task}{\code{TaskManager$add_task()}}
\item \href{#method-get_task}{\code{TaskManager$get_task()}}
\item \href{#method-run_all}{\code{TaskManager$run_all()}}
\item \href{#method-task_run}{\code{TaskManager$task_run()}}
\item \href{#method-run}{\code{TaskManager$run()}}
\item \href{#method-clone}{\code{TaskManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskManager$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_task"></a>}}
\if{latex}{\out{\hypertarget{method-add_task}{}}}
\subsection{Method \code{add_task()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskManager$add_task(task)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_task"></a>}}
\if{latex}{\out{\hypertarget{method-get_task}{}}}
\subsection{Method \code{get_task()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskManager$get_task(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_all"></a>}}
\if{latex}{\out{\hypertarget{method-run_all}{}}}
\subsection{Method \code{run_all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskManager$run_all(log = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-task_run"></a>}}
\if{latex}{\out{\hypertarget{method-task_run}{}}}
\subsection{Method \code{task_run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskManager$task_run(name, log = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskManager$run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
