% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_TaskManager.R
\name{tm_run_task}
\alias{tm_run_task}
\title{Shortcut to run task}
\usage{
tm_run_task(
  task_name,
  index_plan = NULL,
  index_analysis = NULL,
  index_argset = NULL,
  run_as_rstudio_job_loading_from_devtools = FALSE
)
}
\arguments{
\item{task_name}{Name of the task}

\item{index_plan}{Not used}

\item{index_analysis}{Not used}

\item{index_argset}{Deprecated}

\item{run_as_rstudio_job_loading_from_devtools}{Run the task as an rstudio job, loading from devtools}
}
\description{
Shortcut to run task
}
