% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_shortcuts.R
\name{tm_shortcut_argset}
\alias{tm_shortcut_argset}
\title{Shortcut to argset within plan within task}
\usage{
tm_shortcut_argset(
  task_name,
  index_plan = 1,
  index_analysis = 1,
  index_argset = NULL
)
}
\arguments{
\item{task_name}{Name of the task}

\item{index_plan}{Plan within task}

\item{index_analysis}{Not used}

\item{index_argset}{Argset within plan}
}
\description{
Shortcut to argset within plan within task
}
