% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_config_last_updated.r
\name{update_config_last_updated}
\alias{update_config_last_updated}
\title{update_config_last_updated
Updates the config_last_updated db tables}
\usage{
update_config_last_updated(type, tag, date = NULL, datetime = NULL)
}
\arguments{
\item{type}{a}

\item{tag}{a}

\item{date}{Date to set in config_last_updated}

\item{datetime}{Datetime to set in config_last_updated}
}
\description{
update_config_last_updated
Updates the config_last_updated db tables
}
