% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_Schema.R
\name{validator_field_contents_blank}
\alias{validator_field_contents_blank}
\title{Blank data validator}
\usage{
validator_field_contents_blank(data)
}
\arguments{
\item{data}{data passed to schema}
}
\description{
Blank data validator
}
