% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_Schema.R
\name{validator_field_contents_sykdomspulsen}
\alias{validator_field_contents_sykdomspulsen}
\title{validator_field_contents_sykdomspulsen
An example (schema) validator of database data used in Sykdomspulsen}
\usage{
validator_field_contents_sykdomspulsen(data)
}
\arguments{
\item{data}{data passed to schema}
}
\description{
validator_field_contents_sykdomspulsen
An example (schema) validator of database data used in Sykdomspulsen
}
