% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_Schema.R
\name{validator_field_types_blank}
\alias{validator_field_types_blank}
\title{Blank field_types validator}
\usage{
validator_field_types_blank(db_field_types)
}
\arguments{
\item{db_field_types}{db_field_types passed to schema}
}
\description{
Blank field_types validator
}
