% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_Schema.R
\name{validator_field_types_sykdomspulsen}
\alias{validator_field_types_sykdomspulsen}
\title{validator_field_types_sykdomspulsen
An example (schema) validator of field_types used in Sykdomspulsen}
\usage{
validator_field_types_sykdomspulsen(db_field_types)
}
\arguments{
\item{db_field_types}{db_field_types passed to schema}
}
\description{
validator_field_types_sykdomspulsen
An example (schema) validator of field_types used in Sykdomspulsen
}
