% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_folders.R
\name{create_folder_if_doesnt_exist}
\alias{create_folder_if_doesnt_exist}
\title{Creates folder if it doesn't exist}
\usage{
create_folder_if_doesnt_exist(path)
}
\arguments{
\item{path}{The path}
}
\description{
Creates folder if it doesn't exist
}
