% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_tables_last_updated.R
\name{get_config_tables_last_updated}
\alias{get_config_tables_last_updated}
\title{get_config_last_updated
Gets the config_last_updated db table}
\usage{
get_config_tables_last_updated(table_name = NULL)
}
\arguments{
\item{table_name}{Table name}
}
\description{
get_config_last_updated
Gets the config_last_updated db table
}
