% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_tasks_runtime.R
\name{get_config_tasks_stats}
\alias{get_config_tasks_stats}
\title{get_config_last_updated
Gets the config_last_updated db table}
\usage{
get_config_tasks_stats(task = NULL, last_run = FALSE)
}
\arguments{
\item{task}{Task name}

\item{last_run}{Just get the last run?}
}
\description{
get_config_last_updated
Gets the config_last_updated db table
}
