% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_mandatory_db_filter.R
\name{mandatory_db_filter}
\alias{mandatory_db_filter}
\title{mandatory_db_filter}
\usage{
mandatory_db_filter(
  .data,
  granularity_time = NULL,
  granularity_time_not = NULL,
  granularity_geo = NULL,
  granularity_geo_not = NULL,
  country_iso3 = NULL,
  location_code = NULL,
  age = NULL,
  age_not = NULL,
  sex = NULL,
  sex_not = NULL
)
}
\arguments{
\item{.data}{Data}

\item{granularity_time}{Granularity time to filter on (include)}

\item{granularity_time_not}{Granularity time to filter on (exclude)}

\item{granularity_geo}{Granularity geo to filter on (include)}

\item{granularity_geo_not}{Granularity geo to filter on (exclude)}

\item{country_iso3}{country_iso3 to filter on (include)}

\item{location_code}{location_code to filter on (include)}

\item{age}{Age to filter on (include)}

\item{age_not}{Age to filter on (exclude)}

\item{sex}{Sex to filter on (include)}

\item{sex_not}{Sex to filter on (exclude)}
}
\description{
mandatory_db_filter
}
