% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_SurveillanceSystem.R
\name{run_task_sequentially_as_rstudio_job_using_load_all}
\alias{run_task_sequentially_as_rstudio_job_using_load_all}
\title{Run a task sequentially as an RStudio job}
\usage{
run_task_sequentially_as_rstudio_job_using_load_all(
  task_name,
  ss_prefix = "global$ss",
  verbose = FALSE
)
}
\arguments{
\item{task_name}{Task name}

\item{ss_prefix}{The prefix that locates the surveillance system}
}
\description{
Description
}
