% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_SurveillanceSystem.R
\name{SurveillanceSystem_v9}
\alias{SurveillanceSystem_v9}
\title{Surveillance System}
\description{
An R6 Action class contains a function called 'run' that takes three arguments:
\itemize{
\item data (list)
\item arg (list)
\item schema (list)
}
}
\details{
An action is (note: we dont explicitly create these):
\itemize{
\item one R6 Action class
\item A plnr::Plan that provide
a) data (from the plan -- \code{get_data})
b) arguments (from the plan -- \code{get_argset})
to the R6 action class
}

A task is:
\itemize{
\item one R6 Action class
\item A plnr::Plan that provide
a) data (from the plan -- \code{get_data})
b) arguments (from the plan -- \code{get_argset})
to the R6 action class
}

A TaskManager is:
\itemize{
\item one R6 Action class
\item A list of plnr::Plan's
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SurveillanceSystem_v9-new}{\code{SurveillanceSystem_v9$new()}}
\item \href{#method-SurveillanceSystem_v9-add_table}{\code{SurveillanceSystem_v9$add_table()}}
\item \href{#method-SurveillanceSystem_v9-add_task}{\code{SurveillanceSystem_v9$add_task()}}
\item \href{#method-SurveillanceSystem_v9-shortcut_get_task}{\code{SurveillanceSystem_v9$shortcut_get_task()}}
\item \href{#method-SurveillanceSystem_v9-shortcut_get_tables}{\code{SurveillanceSystem_v9$shortcut_get_tables()}}
\item \href{#method-SurveillanceSystem_v9-shortcut_get_argset}{\code{SurveillanceSystem_v9$shortcut_get_argset()}}
\item \href{#method-SurveillanceSystem_v9-shortcut_get_data}{\code{SurveillanceSystem_v9$shortcut_get_data()}}
\item \href{#method-SurveillanceSystem_v9-shortcut_get_plans_argsets_as_dt}{\code{SurveillanceSystem_v9$shortcut_get_plans_argsets_as_dt()}}
\item \href{#method-SurveillanceSystem_v9-shortcut_run_task}{\code{SurveillanceSystem_v9$shortcut_run_task()}}
\item \href{#method-SurveillanceSystem_v9-clone}{\code{SurveillanceSystem_v9$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurveillanceSystem_v9-new"></a>}}
\if{latex}{\out{\hypertarget{method-SurveillanceSystem_v9-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveillanceSystem_v9$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurveillanceSystem_v9-add_table"></a>}}
\if{latex}{\out{\hypertarget{method-SurveillanceSystem_v9-add_table}{}}}
\subsection{Method \code{add_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveillanceSystem_v9$add_table(
  name_access,
  name_grouping = NULL,
  name_variant = NULL,
  field_types,
  keys,
  indexes = NULL,
  validator_field_types = csdb::validator_field_types_blank,
  validator_field_contents = csdb::validator_field_contents_blank
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurveillanceSystem_v9-add_task"></a>}}
\if{latex}{\out{\hypertarget{method-SurveillanceSystem_v9-add_task}{}}}
\subsection{Method \code{add_task()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveillanceSystem_v9$add_task(
  name_grouping = NULL,
  name_action = NULL,
  name_variant = NULL,
  cores = 1,
  permission = NULL,
  plan_analysis_fn_name = NULL,
  for_each_plan = NULL,
  for_each_analysis = NULL,
  universal_argset = NULL,
  upsert_at_end_of_each_plan = FALSE,
  insert_at_end_of_each_plan = FALSE,
  action_fn_name,
  data_selector_fn_name = NULL,
  tables = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name_grouping}}{Name of the task (grouping)}

\item{\code{name_action}}{Name of the task (action)}

\item{\code{name_variant}}{Name of the task (variant)}

\item{\code{cores}}{Number of CPU cores}

\item{\code{permission}}{A permission R6 instance}

\item{\code{plan_analysis_fn_name}}{The name of a function that returns a named list \code{list(for_each_plan = list(), for_each_analysis = NULL)}.}

\item{\code{for_each_plan}}{A list, where each unit corresponds to one data extraction. Generally recommended to use \code{plnr::expand_list}.}

\item{\code{for_each_analysis}}{A list, where each unit corresponds to one analysis within a plan (data extraction). Generally recommended to use \code{plnr::expand_list}.}

\item{\code{universal_argset}}{A list, where these argsets are applied to all analyses univerally}

\item{\code{upsert_at_end_of_each_plan}}{Do you want to upsert your results automatically at the end of each plan?}

\item{\code{insert_at_end_of_each_plan}}{Do you want to insert your results automatically at the end of each plan?}

\item{\code{action_fn_name}}{The name of the function that will be called for each analysis with arguments \code{data}, \code{argset}, \code{schema}}

\item{\code{data_selector_fn_name}}{The name of a function that will be called to obtain the data for each analysis. The function must have the arguments \code{argset}, \code{schema} and must return a named list.}

\item{\code{tables}}{A named list that maps \code{sc9::config$schemas} for use in \code{action_fn_name} and \code{data_selector_fn_name}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurveillanceSystem_v9-shortcut_get_task"></a>}}
\if{latex}{\out{\hypertarget{method-SurveillanceSystem_v9-shortcut_get_task}{}}}
\subsection{Method \code{shortcut_get_task()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveillanceSystem_v9$shortcut_get_task(task_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurveillanceSystem_v9-shortcut_get_tables"></a>}}
\if{latex}{\out{\hypertarget{method-SurveillanceSystem_v9-shortcut_get_tables}{}}}
\subsection{Method \code{shortcut_get_tables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveillanceSystem_v9$shortcut_get_tables(task_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurveillanceSystem_v9-shortcut_get_argset"></a>}}
\if{latex}{\out{\hypertarget{method-SurveillanceSystem_v9-shortcut_get_argset}{}}}
\subsection{Method \code{shortcut_get_argset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveillanceSystem_v9$shortcut_get_argset(
  task_name,
  index_plan = 1,
  index_analysis = 1
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurveillanceSystem_v9-shortcut_get_data"></a>}}
\if{latex}{\out{\hypertarget{method-SurveillanceSystem_v9-shortcut_get_data}{}}}
\subsection{Method \code{shortcut_get_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveillanceSystem_v9$shortcut_get_data(task_name, index_plan = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurveillanceSystem_v9-shortcut_get_plans_argsets_as_dt"></a>}}
\if{latex}{\out{\hypertarget{method-SurveillanceSystem_v9-shortcut_get_plans_argsets_as_dt}{}}}
\subsection{Method \code{shortcut_get_plans_argsets_as_dt()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveillanceSystem_v9$shortcut_get_plans_argsets_as_dt(task_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurveillanceSystem_v9-shortcut_run_task"></a>}}
\if{latex}{\out{\hypertarget{method-SurveillanceSystem_v9-shortcut_run_task}{}}}
\subsection{Method \code{shortcut_run_task()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveillanceSystem_v9$shortcut_run_task(
  task_name,
  run_as_rstudio_job_loading_from_devtools = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SurveillanceSystem_v9-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SurveillanceSystem_v9-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SurveillanceSystem_v9$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
