% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_TaskJob.R
\name{TaskJob}
\alias{TaskJob}
\title{TaskJob R6 Class}
\description{
Run a cs9 task in a background R process with live output streaming back to
the current R console via \code{\link[later:later]{later::later()}} polling. Equivalent in spirit to
\code{\link[=run_task_sequentially_as_rstudio_job_using_load_all]{run_task_sequentially_as_rstudio_job_using_load_all()}} but does not
depend on RStudio's job API (which Positron does not implement).
}
\details{
Methods:
$start()      spawn background process and begin polling
$wait(t)      block this session, draining the event loop, until done
$is_alive()   TRUE while the background process is running
$status()     summary
$tail(n)      last n lines of the log file (snapshot, not live)
$kill()       terminate the background process
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-TaskJob-initialize}{\code{TaskJob$new()}}
    \item \href{#method-TaskJob-start}{\code{TaskJob$start()}}
    \item \href{#method-TaskJob-is_alive}{\code{TaskJob$is_alive()}}
    \item \href{#method-TaskJob-status}{\code{TaskJob$status()}}
    \item \href{#method-TaskJob-wait}{\code{TaskJob$wait()}}
    \item \href{#method-TaskJob-kill}{\code{TaskJob$kill()}}
    \item \href{#method-TaskJob-tail}{\code{TaskJob$tail()}}
    \item \href{#method-TaskJob-clone}{\code{TaskJob$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskJob-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-TaskJob-initialize}{}}}
\subsection{\code{TaskJob$new()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TaskJob$new(
  task_name,
  ss_prefix = "global$ss",
  log_dir = tempdir(check = TRUE)
)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskJob-start"></a>}}
\if{latex}{\out{\hypertarget{method-TaskJob-start}{}}}
\subsection{\code{TaskJob$start()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TaskJob$start()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskJob-is_alive"></a>}}
\if{latex}{\out{\hypertarget{method-TaskJob-is_alive}{}}}
\subsection{\code{TaskJob$is_alive()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TaskJob$is_alive()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskJob-status"></a>}}
\if{latex}{\out{\hypertarget{method-TaskJob-status}{}}}
\subsection{\code{TaskJob$status()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TaskJob$status()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskJob-wait"></a>}}
\if{latex}{\out{\hypertarget{method-TaskJob-wait}{}}}
\subsection{\code{TaskJob$wait()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TaskJob$wait(timeout_s = 600)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskJob-kill"></a>}}
\if{latex}{\out{\hypertarget{method-TaskJob-kill}{}}}
\subsection{\code{TaskJob$kill()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TaskJob$kill()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskJob-tail"></a>}}
\if{latex}{\out{\hypertarget{method-TaskJob-tail}{}}}
\subsection{\code{TaskJob$tail()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TaskJob$tail(n = 20)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskJob-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TaskJob-clone}{}}}
\subsection{\code{TaskJob$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TaskJob$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
